license_portal/requirements.txt
asgiref==3.7.2
Django==3.2
djangorestframework==3.12.2
coreapi==2.3.3
mysqlclient==2.2.0
sqlparse==0.4.4
typing_extensions==4.6.3